Can drop flashlight, mags spawn in piles, etc.
Added magazine spawns in bullet piles. If a mag spawns, no individual
bullets will spawn in the pile. The magazine will spawn with between 0
and capacity bullets in it.

Flashlight can be dropped if weapon is holstered, and no mags are in
hand.
Only one flashlight can be carried at a time - prevent picking up
unwanted flashlights.

Fixed r_10 to adjust the height of the uppermost bathroom furniture (it
was all floating above ground and looked weird.